requirements.txt
flask
openai
tavily-python
trafilatura
pypdf
python-dotenv